[fix] move into data directory
