+++ src/main/java/com/gg/server/domain/noti/data/NotiRepository.java
@@ -1,10 +1,9 @@
-package com.gg.server.domain.noti;
+package com.gg.server.domain.noti.data;
 
 import com.gg.server.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface NotiRepository extends JpaRepository<Noti, Long> {
     List<Noti> findByUser(User user);
