+++ src/main/java/com/gg/server/domain/noti/service/NotiService.java
@@ -1,13 +1,12 @@
 package com.gg.server.domain.noti.service;
 
-import com.gg.server.domain.noti.Noti;
-import com.gg.server.domain.noti.NotiRepository;
+import com.gg.server.domain.noti.data.Noti;
+import com.gg.server.domain.noti.data.NotiRepository;
 import com.gg.server.domain.noti.dto.NotiDto;
 import com.gg.server.domain.user.User;
 import com.gg.server.domain.user.UserRepository;
 import com.gg.server.domain.user.dto.UserDto;
 import lombok.AllArgsConstructor;
-import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
